# Conflicts:
#	wadsrc/static/zscript.txt

+++ src/memarena.h
@@ -40,7 +40,7 @@
 class FMemArena
 {
 public:
-	FMemArena(int blocksize = 10*1024);
+	FMemArena(size_t blocksize = 10*1024);
 	~FMemArena();
 
 	void *Alloc(size_t size);
@@ -55,7 +55,7 @@ protected:
 
 	Block *TopBlock;
 	Block *FreeBlocks;
-	int BlockSize;
+	size_t BlockSize;
 };
 
 // An arena specializing in storage of FStrings. It knows how to free them,
